mcp_python/requirements.txt
mcp>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
python-dotenv